JK - Changed menu to component
